+++ authentication_lib/src/test/java/org/octri/authentication/server/security/service/PasswordResetTokenServiceTest.java
@@ -1,14 +1,11 @@
 package org.octri.authentication.server.security.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.Date;
 
 import org.junit.jupiter.api.BeforeEach;
@@ -53,31 +50,6 @@ public class PasswordResetTokenServiceTest {
 		assertNotNull(token.getExpiryDate(), "There should be an expiration date");
 		assertTrue(token.getExpiryDate().after(new Date()),
 				"The expiration date should be far enough in the future for this test to pass");
-	}
-
-	@Test
-	public void testForValidPasswordResetToken() {
-		PasswordResetToken validToken = new PasswordResetToken(user);
-		when(passwordResetTokenService.findByToken("secret-token")).thenReturn(validToken);
-		assertTrue(passwordResetTokenService.isValidPasswordResetToken("secret-token"), "Token must be valid");
-	}
-
-	@Test
-	public void testForInvalidPasswordResetTokenWhenExpired() {
-		Instant now = Instant.now();
-		PasswordResetToken invalidToken = new PasswordResetToken(user);
-		// Set a date far in the past for this test to pass.
-		invalidToken.setExpiryDate(Date.from(now.minus(1000, ChronoUnit.MINUTES)));
-
-		when(passwordResetTokenService.findByToken("secret-token")).thenReturn(invalidToken);
-		assertFalse(passwordResetTokenService.isValidPasswordResetToken("secret-token"), "Token must not be expired");
-	}
-
-	@Test
-	public void testForInvalidPasswordResetTokenIfDoesNotExist() {
-		when(passwordResetTokenService.findByToken("secret-token")).thenReturn(null);
-		assertFalse(passwordResetTokenService.isValidPasswordResetToken("secret-token"),
-				"Token must exist in the database");
 	}
 
 	@Test